refactor: introduce base entity
