+++ custom_components/satel/binary_sensor.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from homeassistant.components.binary_sensor import BinarySensorEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceInfo
 
 from . import SatelHub
 from .const import DOMAIN
+from .entity import SatelEntity
 
 
 async def async_setup_entry(
@@ -24,23 +24,14 @@ async def async_setup_entry(
     async_add_entities(entities, True)
 
 
-class SatelZoneBinarySensor(BinarySensorEntity):
+class SatelZoneBinarySensor(SatelEntity, BinarySensorEntity):
     """Binary sensor for a Satel zone."""
 
     def __init__(self, hub: SatelHub, zone_id: str, name: str) -> None:
-        self._hub = hub
+        super().__init__(hub)
         self._zone_id = zone_id
         self._attr_name = name
         self._attr_unique_id = f"satel_zone_{zone_id}"
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device information for this entity."""
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._hub.host)},
-            manufacturer="Satel",
-            name="Satel Alarm",
-        )
 
     async def async_update(self) -> None:
         status = await self._hub.send_command(f"ZONE {self._zone_id}")
