+++ custom_components/satel/switch.py
@@ -5,10 +5,10 @@ from __future__ import annotations
 from homeassistant.components.switch import SwitchEntity
 from homeassistant.config_entries import ConfigEntry
 from homeassistant.core import HomeAssistant
-from homeassistant.helpers.device_registry import DeviceInfo
 
 from . import SatelHub
 from .const import DOMAIN
+from .entity import SatelEntity
 
 
 async def async_setup_entry(
@@ -24,24 +24,15 @@ async def async_setup_entry(
     async_add_entities(entities, True)
 
 
-class SatelOutputSwitch(SwitchEntity):
+class SatelOutputSwitch(SatelEntity, SwitchEntity):
     """Switch to control Satel output."""
 
     def __init__(self, hub: SatelHub, output_id: str, name: str) -> None:
-        self._hub = hub
+        super().__init__(hub)
         self._output_id = output_id
         self._attr_name = name
         self._attr_is_on = False
         self._attr_unique_id = f"satel_output_{output_id}"
-
-    @property
-    def device_info(self) -> DeviceInfo:
-        """Return device information for this entity."""
-        return DeviceInfo(
-            identifiers={(DOMAIN, self._hub.host)},
-            manufacturer="Satel",
-            name="Satel Alarm",
-        )
 
     async def async_turn_on(self, **kwargs) -> None:
         await self._hub.send_command(f"OUTPUT {self._output_id} ON")
